Added raytracer, optical system modules & examples
Ray tracer module does the work of representing the physics of an
optical system, i.e. 3d Snells law and the interaction of rays
with spherical surfaces. Optical system module can be used to
specify a particular optical system. This is done in exampleuses.py
where we instantiate a OpticalSystem class member and plot ray
diagrams in 3d.